Add support for repeatable key groups
The variable devil-repeatable-keys is now a list of lists instead of a
list.  Earlier each item in the list represented a repeatable Devil
key sequence.  However, now each item is a list that represents a
group of repeatable Devil key sequences.  When any key sequence in a
group is typed, any key in the same group as the key sequence typed
may be repeated merely by typing the last character of the key.

For those users who customise devil-repeatable-keys, this change
happens to be a breaking change, unfortunately.  Users who do not
customise it will experience no adverse effects.  However, users who
do customise it need to update the value they set
devil-repeatable-keys to from being a list of strings to being a list
of lists of strings.