Revert devil-translations to constructed list
Commit e0a12cd changed the default value of devil-translations to a
quoted list.  However, that change now appears to be misguided because
a user may have existing configuration that modifies this list.
Therefore revert the value of this variable from a quoted list to a
list constructed with the list function.

Thanks to Morgan Willcock for reviewing the earlier commit and
pointing this issue out.