Add option to set devil-mode in all buffers
This changes introduces a customisable variable named
`devil-global-sets-buffer-default` which indicates that enabling the
global minor-mode will also modify the default value of devil-mode.
By doing so, Devil will now be active in all buffers by default, even
for buffers which have not run hooks which could a apply a minor mode.

This fixes the problem of being defaulted into or switching into
buffers where Devil is not active, and finding that Devil key
sequences are not working.

The documentation which has been added emphasises that the use-case for
the new variable is limited and it is much safer to deal with
problematic buffers directly if the user actually notices a problem.