Add optional repeat suppression
This change introduces an option which allows translated key bindings
to avoid being repeated, to be used by Devil or by other packages
where the Devil repeat map should not be activated.

Repeat suppression is configured by maintaining a list of functions
which are treated as repeat suppression predicates.  Interactive
functions will not be repeated when their symbol is a list member.
Non-interactive functions are called with the binding as the argument
and should return a non-nil value when the repeat should be
suppressed.

By default the predicate list contains the symbol
devil--isearch-suppress-repeat-p which suppresses the repeat of
isearch-forward while the isearch specific mode map is not active.
This fixes initiating an isearch with the default list of repeatable
keys and trying to enter the letter 's' as the search criteria.